Add open escape option
Removes alternate states for the escape so that the whole world is still accessible. This doesn't include logic changes and doesn't affect the escape timer. There's also likely issues with the timer display if you enter a room with too many enemies. map rando removes enemies during the escape which is something I can implement later.